src/app/panel/panel.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { MatExpansionModule } from '@angular/material/expansion';
import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
import { PanelComponent } from './panel.component';

describe('PanelComponent', () => {
  let component: PanelComponent;
  let fixture: ComponentFixture<PanelComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [PanelComponent],
      imports: [MatExpansionModule, BrowserAnimationsModule]
    }).compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(PanelComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create a panel', () => {
    expect(component).toBeTruthy();
  });

  it(`should have as title 'List of goods'`, () => {
    expect(component.panelTitle).toEqual('List of goods');
 });

  it('should render title for the panel', () => {
    const fixture = TestBed.createComponent(PanelComponent);
    fixture.detectChanges();
    const compiled = fixture.nativeElement;
    expect(compiled.querySelector('.panel h1').textContent).toContain('List of goods!');
  });
});
